Add files via upload
A solution of google codejam qualtification round problem. A problem in which + is upside pancake and - is downside pancake. We have to  find wheather we could make all pancakes upside and if possible how many steps it would require. For example:
After space a number is given it would say that at a time this number of pancakes we can flip.
---+-++-  3  = 3
+++++ 4  = 0
-+-+-  4  = impossible